requirements.txt
# LangChain + Google Generative AI
langchain==0.1.12
langchain-google-genai==0.0.11
google-generativeai>=0.4.1,<0.5.0

# Graph database
neo4j==5.17.0


# Data handling
pandas==2.2.1
numpy==1.26.4

# Typing and environment support
python-dotenv==1.0.1
typing-extensions==4.9.0

